+++ lib/wip/auth.rb
@@ -1,7 +1,7 @@
-require "wip"
-
-class Wip::Auth
-  def self.api_key
-    ENV['WIP_API_KEY']
+module Wip
+  class Auth
+    def self.api_key
+      ENV['WIP_API_KEY']
+    end
   end
 end