+++ lib/wip/user.rb
@@ -3,32 +3,49 @@ require "wip"
 require "wip/client"
 
 class Wip::User
-  attr_accessor :avatar_url, :best_streak, :completed_todos_count, :first_name, :id, :last_name, :streak, :streaking, :time_zone, :url, :username
+  attr_accessor :avatar_url, :best_streak, :completed_todos_count, :first_name, :id, :last_name, :streak, :streaking, :time_zone, :url, :username, :todos
+
+  def self.default_selection
+    %{
+      avatar_url
+      best_streak
+      completed_todos_count
+      first_name
+      id
+      last_name
+      streak
+      streaking
+      time_zone
+      todos {#{Wip::Todo.default_selection}}
+      url
+      username
+    }
+  end
 
   def self.viewer
     client = Wip::Client.new
     find_query = %{
       {
-        viewer {
-          avatar_url
-          best_streak
-          completed_todos_count
-          first_name
-          id
-          last_name
-          streak
-          streaking
-          time_zone
-          url
-          username
-        }
+        viewer {#{default_selection}}
       }
     }
     client.request find_query
     parse client.data("viewer")
   end
 
-  def initialize(avatar_url: nil, best_streak: nil, completed_todos_count: nil, first_name: nil, id: nil, last_name: nil, streak: nil, streaking: nil, time_zone: nil, url: nil, username: nil)
+  def self.find(id)
+    client = Wip::Client.new
+
+    find_query = %{
+      {
+        user(id: #{id}) {#{default_selection}}
+      }
+    }
+    client.request find_query
+    parse client.data("user")
+  end
+
+  def initialize(avatar_url: nil, best_streak: nil, completed_todos_count: nil, first_name: nil, id: nil, last_name: nil, streak: nil, streaking: nil, time_zone: nil, url: nil, username: nil, todos: [])
     @avatar_url = avatar_url
     @best_streak = best_streak
     @completed_todos_count = completed_todos_count
@@ -40,12 +57,21 @@ class Wip::User
     @time_zone = time_zone
     @url = url
     @username = username
+    @todos = todos
   end
 
   def self.parse(data)
     new.tap do |user|
-      data.each do |key, value|
-        user.send("#{key}=", key == "id" ? value.to_i : value)
+      data.each do |key, raw_value|
+        value = case key
+        when "id"
+          raw_value.to_i
+        when "todos"
+          raw_value.collect { |v| Wip::Todo.parse v }
+        else
+          raw_value
+        end
+        user.send("#{key}=", value)
       end
     end
   end