+++ lib/wip/cli.rb
@@ -25,12 +25,14 @@ class Wip::CLI < Thor
   end
 
   desc "todos", "List viewer todos"
-  method_option :completed, type: :boolean, aliases: '-c', default: nil
-  method_option :filter, type: :string, aliases: '-f', default: nil
+  method_option :completed, type: :boolean, aliases: '-c'
+  method_option :filter, type: :string, aliases: '-f'
   method_option :limit, type: :numeric, aliases: '-l', default: 5
   method_option :order, type: :string, aliases: '-o', default: "completed_at:desc"
+  method_option :username, type: :string, aliases: '-u'
   def todos
-    user = Wip::User.viewer(todos: options.slice("completed", "filter", "limit", "order"))
+    todos_options = options.slice("completed", "filter", "limit", "order")
+    user = options.username.nil? ? Wip::User.viewer(todos: todos_options) : Wip::User.find(username: options.username, todos: todos_options)
     user.todos.each do |todo|
       puts todo.description
     end